feat: add useDirtyFieldStore zustand store for dirty state tracking
