+++ frontend/src/features/admin-form/create/builder-and-design/BuilderAndDesignDrawer/CreatePageDrawerCloseButton.tsx
@@ -3,10 +3,10 @@ import { BiX } from 'react-icons/bi'
 import { CloseButton } from '@chakra-ui/react'
 
 import { useCreatePageSidebar } from '../../common/CreatePageSidebarContext'
-import { isDirtySelector, useFieldBuilderStore } from '../useFieldBuilderStore'
+import { isDirtySelector, useDirtyFieldStore } from '../useDirtyFieldStore'
 
 export const CreatePageDrawerCloseButton = (): JSX.Element => {
-  const isDirty = useFieldBuilderStore(isDirtySelector)
+  const isDirty = useDirtyFieldStore(isDirtySelector)
   const { handleClose } = useCreatePageSidebar()
 
   const handleCloseDrawer = useCallback(() => {
